rename bundle_rec to recognition; add recognition specific tests folder; other minor changes
