+++ AFQ/recognition/criteria.py
