+++ AFQ/recognition/curvature.py
