+++ AFQ/recognition/roi.py
